[WiP] Finalizing first correct version; uncertainty in aprime
Finalizing first correct version. There is a bit of uncertainty
surrounding computation of aprime. The momentum balance has
density in the denominator in some slides but not always.

Moreover, the implementation of Prandtl's correction for heavily
loaded streamtubes is uncertain since it was shown that a>0.5,
actual CT diverges but the equations shows that this is only
corrected for if a > 0.88 or so.