Finish first working version BEM code for assignment
Tunnel vision resulted that I didn't see the Jupyter notebook
of the prof. with the majority of the code. I compared and the
equations and approach are the same (except computation of loads).
The angle computations are a bit weird though. Nonetheless, the
equidistant version works. Cosine spacing still seems to give
erroneous results.