Vectorize BEM solver to prepare for skewed wake integration
BEM needs to be vectorized to speed-up computation of skewed
wake. In case of a yaw angle, each element in the annulus
experiences a different induced velocity and hence the loading
will be different for each element in the annulu

Vectorize BEM solver to prepare for skewed wake integration

BEM needs to be vectorized to speed-up computation of skewed
wake. In case of a yaw angle, each element in the annulus
experiences a different induced velocity and hence the loading
will be different for each element in the annuli