Change order of computation BEMSolver and add changes flowchart
Changed order of computation in BEMSolver so it better reflects
the loop in the flow chart which makes more sense. Found an
uncertainty in the a' computation: include ftotal in kappa'
or compute a' and then divide by ftotal? It seems that the
differences are negligible.

Flowchart suggestions updated; red-ish indicate remove, blue-ish
indicates add.