Rotate figures 90 degrees due to definition axes BEM Solver
The azimuthal angle is defined to be 0 deg in the North, looking
downstream at the wind turbine. This means that the mathematical cosine
and sine functions must be offset by 90 degrees. Another way to do this
is by switching the cosine and sine definitions for x and y. The latter was done.